fix(webserver): mac webserver support
